move gql and test to appropriate directory
